(feat) merge gateway status monitor with http client
